Split AppStoreConnect tests into smaller files
